Revamped SymbolicLink Creation on Windows 🛠️ - Requires Dev/Admin Privileges
Given that we are once again utilizing the cmd mklink to create symlinks under Windows, it's imperative to execute this script with Developer mode or Administrator privileges. Furthermore, I've expanded and strengthened the privilege checks to ensure a smooth and secure execution. Need to be tested Windows. Expanded the Verbose out to invoke-expressions. Add the ability to Force to Recreate SymLinks. Add error handling of dependencies file